third_party/WebKit/LayoutTests/inspector/sass/test-mapping-good.html

<html>
<head>

<link rel="stylesheet" href="resources/test-mapping-good.css">

<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/debugger-test.js"></script>
<script src="./sass-test.js"></script>
<script>

function test()
{
    var header = InspectorTest.cssModel.styleSheetHeaders().find(header => !!header.sourceMapURL)
    InspectorTest.loadSourceMap(header, onSourceMapLoaded);
    InspectorTest.waitForScriptSource("test-mapping-good.css", onCSSLoaded);
    InspectorTest.waitForScriptSource("test-mapping-good.scss", onSCSSLoaded);

    var cssAST;
    var sassModels = new Map();
    var sourceMap;
    var mapping;

    function onCSSLoaded(uiSourceCode)
    {
        uiSourceCode.requestContentPromise()
            .then(text => InspectorTest.parseCSS(uiSourceCode.originURL(), text))
            .then(ast => cssAST = ast)
            .then(maybeCreateMapping);
    }

    function onSCSSLoaded(uiSourceCode)
    {
        uiSourceCode.requestContentPromise()
            .then(text => InspectorTest.parseSCSS(uiSourceCode.originURL(), text))
            .then(ast => sassModels.set(ast.document.url, ast))
            .then(maybeCreateMapping);
    }

    function onSourceMapLoaded(sm)
    {
        sourceMap = sm;
        maybeCreateMapping();
    }

    function maybeCreateMapping()
    {
        if (!cssAST || !sassModels.size || !sourceMap)
            return;
        mapping = WebInspector.SASSLiveSourceMap.CSSToSASSMapping.fromSourceMap(sourceMap, cssAST, sassModels);
        if (!mapping.isValid()) {
            InspectorTest.addResult("ERROR: mapping is not valid.");
            InspectorTest.completeTest();
            return;
        }
        InspectorTest.validateMapping(mapping, cssAST, sassModels);
        InspectorTest.runTestSuite(testSuite);
    }

    var testSuite = [
        function testCSSRebase(next)
        {
            var cssClone = cssAST.clone();
            cssClone.rules[0].properties[1].remove();
            var cssDiff = WebInspector.SASSSupport.diffModels(cssAST, cssClone);
            var newMapping = mapping.rebaseForCSSDiff(cssDiff);
            if (!newMapping.isValid()) {
                InspectorTest.addResult("ERROR: mapping is not valid.");
                InspectorTest.completeTest();
                return;
            }
            InspectorTest.validateMapping(newMapping, cssClone, sassModels);
            next();
        },

        function testSCSSRebase(next)
        {
            var sassAST = sassModels.valuesArray()[0];
            var sassClone = sassAST.clone();
            sassClone.rules[1].properties[2].remove();
            sassClone.rules[1].properties[1].remove();
            var sassDiff = WebInspector.SASSSupport.diffModels(sassAST, sassClone);
            var newMapping = mapping.rebaseForSASSDiff(sassDiff);
            if (!newMapping.isValid()) {
                InspectorTest.addResult("ERROR: mapping is not valid.");
                InspectorTest.completeTest();
                return;
            }
            var sassModelsClone = new Map();
            sassModelsClone.set(sassClone.document.url, sassClone);
            InspectorTest.validateMapping(newMapping, cssAST, sassModelsClone);
            next();
        },
    ];
}

</script>

</head>

<body onload="runTest()">
<p>Verify CSSToSASSMapping rebaselines.</p>
</body>
</html>
