third_party/WebKit/LayoutTests/inspector/sass/test-mapping-good-expected.txt

Verify CSSToSASSMapping rebaselines.

Mapped CSS: 12
Mapped SCSS: 12
No stale entries found.

Running: testCSSRebase
Mapped CSS: 10
Mapped SCSS: 10
No stale entries found.

Running: testSCSSRebase
Mapped CSS: 8
Mapped SCSS: 8
No stale entries found.
